+++ SQL/New_patches/2021-01-31-renaming_permissions.sql
@@ -1,3 +1,21 @@
+ALTER TABLE permissions
+    ADD COLUMN `moduleID` int(11) unsigned AFTER `description`,
+    ADD COLUMN `action` enum (
+    'View',
+    'Create',
+    'Edit',
+    'Download',
+    'Upload',
+    'Delete',
+    'View/Create',
+    'View/Edit',
+    'View/Download',
+    'View/Upload',
+    'View/Create/Edit',
+    'Create/Edit',
+    'Edit/Upload',
+    'Edit/Upload/Delete') AFTER `moduleID`;
+
 UPDATE permissions SET description='User Accounts - Own Sites', moduleID=(SELECT ID FROM modules WHERE Name='user_accounts'), action='View/Create/Edit' WHERE code='user_accounts';
 UPDATE permissions SET description='User Accounts - All Sites', moduleID=(SELECT ID FROM modules WHERE Name='user_accounts'), action='View/Create/Edit' WHERE code='user_accounts_multisite';
 UPDATE permissions SET description='Help documentation', moduleID=(SELECT ID FROM modules WHERE Name='help_editor'), action='Edit' WHERE code='context_help';
@@ -57,4 +75,5 @@ UPDATE permissions SET description='Documents', moduleID=(SELECT ID FROM modules
 UPDATE permissions SET description='Candidate Surveys', moduleID=(SELECT ID FROM modules WHERE Name='survey_accounts'), action='View' WHERE code='survey_accounts_view';
 UPDATE permissions SET description='Flagged Imaging Entries', moduleID=(SELECT ID FROM modules WHERE Name='imaging_qc'), action='View' WHERE code='imaging_quality_control_view';
 UPDATE permissions SET description='Flagged Behavioural Entries', moduleID=(SELECT ID FROM modules WHERE Name='behavioural_qc'), action='View' WHERE code='behavioural_quality_control_view';
+UPDATE permissions SET description='LORIS API Manual', moduleID=(SELECT ID FROM modules WHERE Name='api_docs'), action='View' WHERE code='api_docs';