+++ modules/issue_tracker/jsx/IssueForm.js
@@ -242,6 +242,7 @@ class IssueForm extends Component {
             label='Assignee'
             emptyOption={true}
             options={this.state.Data.assignees}
+            disabledOptions={this.state.Data.inactiveUsers}
             onUserInput={this.setFormData}
             disabled={!hasEditPermission}
             value={this.state.formData.assignee}
@@ -370,6 +371,10 @@ class IssueForm extends Component {
           // the user's sites.
           if (newIssue) {
             formData.centerID = null;
+            Object.keys(data.inactiveUsers).map((user) => {
+              delete data.assignees[user];
+            });
+            data.inactiveUsers = {};
           } else {
             // if we edit an issue
             // a NULL centerID (= All Sites) is converted to the ALL Sites option