+++ test/test_instrument/test_instrumentTest.php
@@ -60,10 +60,11 @@ class TestInstrumentTestIntegrationTest extends LorisIntegrationTest
         $this->DB->insert(
             'flag',
             [
-                'ID'        => '999999',
-                'SessionID' => '999999',
-                'Test_name' => 'testtest',
-                'CommentID' => '11111111111111111',
+                'ID'         => '999999',
+                'SessionID'  => '999999',
+                'Data_entry' => 'In Progress',
+                'Test_name'  => 'testtest',
+                'CommentID'  => '11111111111111111',
             ]
         );
         // Set up database wrapper and config
@@ -143,7 +144,7 @@ class TestInstrumentTestIntegrationTest extends LorisIntegrationTest
     function testTextElement()
     {
         $this->_landing();
-        $textElement = $this->webDriver->findElement(
+        $this->webDriver->findElement(
             WebDriverBy::Name("testText")
         )->sendKeys("Test Text successful");
         $this->webDriver->findElement(
@@ -164,7 +165,7 @@ class TestInstrumentTestIntegrationTest extends LorisIntegrationTest
     function testCheckBoxElement()
     {
         $this->_landing();
-        $textElement = $this->webDriver->findElement(
+        $this->webDriver->findElement(
             WebDriverBy::Name("testCheckbox")
         )->click();
         $this->webDriver->findElement(
@@ -186,8 +187,11 @@ class TestInstrumentTestIntegrationTest extends LorisIntegrationTest
     {
         // select 'Yes' option and check it.
         $this->_landing();
-        $select  = $this->safeFindElement(WebDriverBy::Name("consent"));
+        $select  = $this->safeFindElement(
+            WebDriverBy::cssSelector("select[name='consent']")
+        );
         $element = new WebDriverSelect($select);
+        sleep(1);
         $element->selectByVisibleText("Yes");
 
         $this->webDriver->findElement(
@@ -202,9 +206,13 @@ class TestInstrumentTestIntegrationTest extends LorisIntegrationTest
 
         // select 'No' option and check it.
         $this->_landing();
-        $select  = $this->safeFindElement(WebDriverBy::Name("consent"));
+        $select  = $this->safeFindElement(
+            WebDriverBy::cssSelector("select[name='consent']")
+        );
         $element = new WebDriverSelect($select);
         $element->selectByVisibleText("No");
+
+        sleep(1);
 
         $this->webDriver->findElement(
             WebDriverBy::Name("fire_away")