+++ webpack.config.js
@@ -41,10 +41,10 @@ const resolve = {
     Tabs: path.resolve(__dirname, './jsx/Tabs'),
     TriggerableModal: path.resolve(__dirname, './jsx/TriggerableModal'),
     Card: path.resolve(__dirname, './jsx/Card'),
-    Password: path.resolve(__dirname, './jsx/Password'),
   },
   extensions: ['*', '.js', '.jsx', '.json'],
 };
+
 const mod = {
   rules: [
     {
@@ -140,7 +140,6 @@ const config = [
       Form: './jsx/Form.js',
       Markdown: './jsx/Markdown.js',
       CSSGrid: './jsx/CSSGrid.js',
-      Password: './jsx/Password.js',
     },
     output: {
       path: __dirname + '/htdocs/js/components/',
@@ -208,6 +207,7 @@ const config = [
     'index',
     'CandidateIssuesWidget',
   ]),
+  lorisModule('login', ['loginIndex']),
   lorisModule('publication', ['publicationIndex', 'viewProjectIndex']),
   lorisModule('document_repository', ['docIndex', 'editFormIndex']),
   lorisModule('candidate_parameters', [
@@ -241,6 +241,7 @@ const config = [
     'react.sidebar',
     'react.tabs',
   ]),
+  lorisModule('dictionary', ['dataDictIndex']),
   lorisModule('dqt', [
     'components/expansionpanels',
     'components/searchabledropdown',
@@ -289,6 +290,7 @@ const config = [
   lorisModule('server_processes_manager', ['server_processes_managerIndex']),
   lorisModule('instruments', ['CandidateInstrumentList']),
   lorisModule('candidate_profile', ['CandidateInfo']),
+  lorisModule('api_docs', ['swagger-ui_custom']),
 ];
 
 // Support project overrides